requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
pyyaml>=6.0.1
tqdm>=4.66.0
scipy>=1.12.0
